+++ lib/signalman/mail_handler.rb
@@ -1,0 +1,6 @@
+require "signalman/base_handler"
+
+module Signalman
+  class MailHandler < BaseHandler
+  end
+end
