+++ lib/signalman/view_handler.rb
@@ -1,0 +1,9 @@
+require "signalman/base_handler"
+
+module Signalman
+  class ViewHandler < BaseHandler
+    def skip?
+      event.payload[:identifier].include?("app/views/signalman/")
+    end
+  end
+end
